makeArc Changes
in this change, makeArc is now a static method, so I got rid of the need of instancing the TFMathExtension class and replaced the object declaration with a static one. I've also added a two new variables, arrayX and arrayY, to test points in all four quadrants of the coordinate plane. But, you'll need to add code to increment the indices as I am just editing the raw file without android studio.